+++ packages/describe-package/test/describe-project.test.ts
@@ -3,22 +3,23 @@ import { setupProject } from './helpers';
 
 import describeProject from '../src/describe-project';
 
-let fns;
+let members;
 
 // Load the fixture once and then run a bunch of tests against it
 test.before(async () => {
   const project = await setupProject('stroopwafel');
-  fns = await describeProject(project);
+  members = await describeProject(project);
 });
 
-const get = (name) => fns.find((fn) => fn.name === name);
+const get = (name) => members.find((fn) => fn.name === name);
 
-test('List all the exported functions', async (t) => {
-  t.assert(fns.length === 4);
+test('List all the exported members', async (t) => {
+  t.assert(members.length === 5);
   t.truthy(get('traditional'));
   t.truthy(get('oneFlavour'));
   t.truthy(get('manyFlavours'));
   t.truthy(get('fn'));
+  t.truthy(get('flavours'));
 });
 
 test('Does not include private functions', async (t) => {
@@ -69,4 +70,11 @@ test('Parse an empty file', async (t) => {
   const project = await setupProject('empty');
   const fns = await describeProject(project);
   t.is(fns.length, 0);
+});
+
+test('Recognise a namespace', async (t) => {
+  const ns = get('flavours');
+  t.is(ns.type, 'namespace')
+
+  // Note that we don't do a lot with the namespace right now - we just acknowledge that its there
 });