+++ packages/describe-package/test/fixtures/stroopwafel-flavours.d.ts
@@ -1,0 +1,3 @@
+export const strawberry = 'strawberry';
+
+export const chocolate = 'chocolate';