+++ integration-tests/worker/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @openfn/integration-tests-worker
 
+## 1.0.39
+
+### Patch Changes
+
+- Updated dependencies [7ddc5d8]
+- Updated dependencies [4deb5d4]
+  - @openfn/ws-worker@1.1.4
+  - @openfn/engine-multi@1.1.4
+  - @openfn/lightning-mock@2.0.4
+
 ## 1.0.38
 
 ### Patch Changes