+++ packages/lightning-mock/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/lightning-mock
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [cecdb60]
+  - @openfn/runtime@1.1.2
+  - @openfn/engine-multi@1.1.4
+
 ## 2.0.3
 
 ### Patch Changes